refactor(master): move AGV update structs into master/updates/
